pyproject.toml
[project]
name = "chromadb-rag"
version = "0.1.0"
description = "Simple rag with ChromaDB as vector database with Google free LLMs for retrieval and embedding"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "google-genai>=1.0.0",
    "python-dotenv>=1.2.2",
    "markitdown[all]>=0.1.0",
    "tiktoken>=0.8.0",
    "rich>=13.0.0",
]
